move getCatalogs
